More clean up
Former-commit-id: 67e87bef8225493aba1ce831d6b945485b975ef0 [formerly 3f05c5fea23bc9dab4fa4f0944b07bf21e0a1848] [formerly 4f70927078076265e7f274544fc53c4f6556fc53 [formerly 0f6ed9981401503b1f3561ccdcde02ce258f1d15] [formerly 2113e00762f556b303c687aca1caa1a1e8775db4] [formerly b4db7344ca06c36a237208dae8bd3b0d7880b018 [formerly de25664c15bec1bbf4c4396e82bf0089f992421a]] [formerly 4f70927078076265e7f274544fc53c4f6556fc53 [formerly 0f6ed9981401503b1f3561ccdcde02ce258f1d15] [formerly 2113e00762f556b303c687aca1caa1a1e8775db4] [formerly b4db7344ca06c36a237208dae8bd3b0d7880b018 [formerly de25664c15bec1bbf4c4396e82bf0089f992421a]] [formerly b2ca1825590115767bd958f9ab327a4806fb4a92]]]
Former-commit-id: 598b58756924e41fdc12d4e09acbecb8d2e50551
Former-commit-id: 8ab7f5c27962b320c272d0393b7ac916ad8d08d3
Former-commit-id: 26f278d137c6c3d7dc3c5e4fc0ad55fd9847052a
Former-commit-id: 4035d440b05ecf157e4b60c7c5be06c358968e98
Former-commit-id: 1c603cb52844c4eefcb7e7fce06d509b409e7617 [formerly 8d32f7b98f18718a42c31efd3977dcf2b2e41054]
Former-commit-id: c5958b024042c221515cac41a3596f9c4e22c5ba
Former-commit-id: 6abb446e2c66d5ff8a7a75a22964578eea32101c

+++ Source/RawParser/Model/Parser/Decoder/Decompressor/BitPump/BitPumpPlain.cs
@@ -89,7 +89,7 @@ namespace RawNet.Decoder.Decompressor
         {
             fixed (byte* t = &buffer[off >> 3])
             {
-                var v = (ushort)(*(int*)t >> ((int)off & 7) & ((1 << nbits) - 1));
+                var v = (ushort)(*(int*)t >> (off & 7) & ((1 << nbits) - 1));
                 off += nbits;
                 return v;
             }
